refactor: combined ApplicationsApproved,-Received,-Rejected to -Overview
fix: removed unused import
