+++ packages/round-manager/src/features/round/ApplicationsOverview.tsx
@@ -1,11 +1,8 @@
-import { useEffect, useState } from "react";
-import { Link, useParams } from "react-router-dom";
-import { InboxInIcon as NoApplicationsForRoundIcon } from "@heroicons/react/outline";
+import { Link } from "react-router-dom";
 import { useWallet } from "../common/Auth";
 import { Spinner } from "../common/Spinner";
 import {
   BasicCard,
-  Button,
   CardContent,
   CardDescription,
   CardHeader,
@@ -19,32 +16,42 @@ import {
   ProgressStep,
   ProjectStatus,
 } from "../api/types";
+import { useEffect, useState } from "react";
 import ConfirmationModal from "../common/ConfirmationModal";
 import {
   AdditionalGasFeesNote,
   ApplicationHeader,
-  ApprovedApplicationsCount,
   Cancel,
+  Continue,
+  ApprovedApplicationsCount,
   RejectedApplicationsCount,
   Select,
+  NoApplicationsContent,
 } from "./BulkApplicationCommon";
-import { useApplicationByRoundId } from "../../context/application/ApplicationContext";
 import { datadogLogs } from "@datadog/browser-logs";
 import { useBulkUpdateGrantApplications } from "../../context/application/BulkUpdateGrantApplicationContext";
 import ProgressModal from "../common/ProgressModal";
 
-export default function ApplicationsReceived() {
-  const { id } = useParams();
+export default function ApplicationsOverview({
+  id,
+  applications,
+  isLoading,
+  applicationStatus,
+}: {
+  id: string | undefined;
+  applications: GrantApplication[] | undefined;
+  isLoading: boolean;
+  applicationStatus: ApplicationStatus;
+}) {
   const { chain } = useWallet();
-  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-  const { applications, isLoading } = useApplicationByRoundId(id!);
-  const pendingApplications =
+
+  const filteredApplications =
     applications?.filter(
-      (a) => a.status == ApplicationStatus.PENDING.toString()
+      (a: GrantApplication) => a.status === applicationStatus.toString()
     ) || [];
 
   const [bulkSelect, setBulkSelect] = useState(false);
-  const [openModal, setOpenModal] = useState(false);
+  const [openConfirmationModal, setOpenConfirmationModal] = useState(false);
   const [openProgressModal, setOpenProgressModal] = useState(false);
   const [selected, setSelected] = useState<GrantApplication[]>([]);
 
@@ -88,7 +95,7 @@ export default function ApplicationsReceived() {
   useEffect(() => {
     if (!isLoading || !bulkSelect) {
       setSelected(
-        (pendingApplications || []).map((application) => {
+        (filteredApplications || []).map((application) => {
           return {
             id: application.id,
             round: application.round,
@@ -101,10 +108,14 @@ export default function ApplicationsReceived() {
     }
   }, [applications, isLoading, bulkSelect]); // eslint-disable-line react-hooks/exhaustive-deps
 
-  const toggleSelection = (id: string, status: ProjectStatus) => {
+  const toggleSelection = (
+    id: string,
+    status1: ProjectStatus,
+    status2: ProjectStatus
+  ) => {
     const newState = selected?.map((grantApp: GrantApplication) => {
       if (grantApp.id === id) {
-        const newStatus = grantApp.status === status ? "PENDING" : status;
+        const newStatus = grantApp.status === status1 ? status2 : status1;
         return { ...grantApp, status: newStatus };
       }
 
@@ -122,13 +133,11 @@ export default function ApplicationsReceived() {
   const handleBulkReview = async () => {
     try {
       setOpenProgressModal(true);
-      setOpenModal(false);
+      setOpenConfirmationModal(false);
       await bulkUpdateGrantApplications({
         // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
         roundId: id!,
-        applications: selected.filter(
-          (application) => application.status !== "PENDING"
-        ),
+        applications: updateFilter(applicationStatus.toString(), selected),
       });
       setBulkSelect(false);
       setOpenProgressModal(false);
@@ -140,8 +149,8 @@ export default function ApplicationsReceived() {
   };
 
   return (
-    <div>
-      {pendingApplications && pendingApplications.length > 0 && (
+    <>
+      {filteredApplications && filteredApplications.length > 0 && (
         <div className="flex items-center justify-end mb-4">
           <span className="text-grey-400 text-sm mr-6">
             Save in gas fees by approving/rejecting multiple applications at
@@ -156,7 +165,7 @@ export default function ApplicationsReceived() {
       )}
       <CardsContainer>
         {!isLoading &&
-          pendingApplications?.map((application, index) => (
+          filteredApplications?.map((application, index) => (
             <BasicCard
               key={index}
               className="application-card"
@@ -166,11 +175,31 @@ export default function ApplicationsReceived() {
                 <ApplicationHeader
                   bulkSelect={bulkSelect}
                   applicationStatus={checkSelectionStatus(application.id)}
-                  approveOnClick={() =>
-                    toggleSelection(application.id, "APPROVED")
+                  approveOnClick={
+                    applicationStatus === "PENDING"
+                      ? () =>
+                          toggleSelection(application.id, "APPROVED", "PENDING")
+                      : applicationStatus === "REJECTED"
+                      ? () =>
+                          toggleSelection(
+                            application.id,
+                            "REJECTED",
+                            "APPROVED"
+                          )
+                      : undefined
                   }
-                  rejectOnClick={() =>
-                    toggleSelection(application.id, "REJECTED")
+                  rejectOnClick={
+                    applicationStatus === "PENDING"
+                      ? () =>
+                          toggleSelection(application.id, "REJECTED", "PENDING")
+                      : applicationStatus === "APPROVED"
+                      ? () =>
+                          toggleSelection(
+                            application.id,
+                            "APPROVED",
+                            "REJECTED"
+                          )
+                      : undefined
                   }
                   application={application}
                 />
@@ -188,106 +217,71 @@ export default function ApplicationsReceived() {
         {isLoading && (
           <Spinner text="We're fetching your Grant Applications." />
         )}
-        {!isLoading && pendingApplications?.length === 0 && (
+        {!isLoading && filteredApplications?.length === 0 && (
           <NoApplicationsContent />
         )}
       </CardsContainer>
       {selected &&
-        selected?.filter((obj) => obj.status !== "PENDING").length > 0 && (
-          <>
-            <div className="fixed w-full left-0 bottom-0 bg-white">
-              <hr />
-              <div className="flex justify-end items-center py-5 pr-20">
-                <NumberOfApplicationsSelectedMessage
-                  grantApplications={selected}
-                  predicate={(selected) => selected.status !== "PENDING"}
-                />
-                <Continue onClick={() => setOpenModal(true)} />
-              </div>
-            </div>
-            <ConfirmationModal
-              title={"Confirm Decision"}
-              confirmButtonText={
-                isBulkUpdateLoading ? "Confirming..." : "Confirm"
-              }
-              body={
-                <>
-                  <p className="text-sm text-grey-400">
-                    {
-                      "You have selected multiple Grant Applications to approve and/or reject."
-                    }
-                  </p>
-                  <div className="flex my-8 gap-16 justify-center items-center text-center">
-                    <ApprovedApplicationsCount grantApplications={selected} />
-                    <span className="text-4xl font-thin">|</span>
-                    <RejectedApplicationsCount grantApplications={selected} />
-                  </div>
-                  <AdditionalGasFeesNote />
-                </>
-              }
-              confirmButtonAction={handleBulkReview}
-              cancelButtonAction={() => setOpenModal(false)}
-              isOpen={openModal}
-              setIsOpen={setOpenModal}
-            />
-          </>
+        updateFilter(applicationStatus.toString(), selected).length > 0 && (
+          <Continue
+            grantApplicationLength={
+              updateFilter(applicationStatus.toString(), selected).length
+            }
+            onClick={() => setOpenConfirmationModal(true)}
+          />
         )}
+
+      <ConfirmationModal
+        title={"Confirm Decision"}
+        confirmButtonText={isBulkUpdateLoading ? "Confirming..." : "Confirm"}
+        confirmButtonAction={handleBulkReview}
+        cancelButtonAction={() => setOpenConfirmationModal(false)}
+        body={
+          <>
+            <p className="text-sm text-grey-400">
+              {applicationStatus.toString() == "REJECTED"
+                ? "You have selected multiple Grant Applications to be approved."
+                : applicationStatus.toString() == "APPROVED"
+                ? "You have selected multiple Grant Applications to be rejected."
+                : "You have selected multiple Grant Applications to approve and/or reject."}
+            </p>
+            <div className="flex my-8 gap-16 justify-center items-center text-center">
+              {applicationStatus.toString() !== "APPROVED" && (
+                <ApprovedApplicationsCount grantApplications={selected} />
+              )}
+              {applicationStatus.toString() === "PENDING" && (
+                <span className="text-4xl font-thin">|</span>
+              )}
+              {applicationStatus.toString() !== "REJECTED" && (
+                <RejectedApplicationsCount grantApplications={selected} />
+              )}
+            </div>
+            <AdditionalGasFeesNote />
+          </>
+        }
+        isOpen={openConfirmationModal}
+        setIsOpen={setOpenConfirmationModal}
+      />
       <ProgressModal
         isOpen={openProgressModal}
         subheading={"Please hold while we update the grant applications."}
         steps={progressSteps}
       />
-    </div>
-  );
-}
-
-function NoApplicationsMessage() {
-  return (
-    <>
-      <h2 className="mt-8 text-2xl antialiased">No Applications</h2>
-      <div className="mt-2 text-sm">
-        Applications have not been submitted yet.
-      </div>
-      <div className="text-sm">
-        Try promoting your Grant Program to get more traction!
-      </div>
     </>
   );
 }
 
-function NoApplicationsContent() {
-  return (
-    <div className="flex flex-center flex-col mx-auto h-screen items-center text-center mt-32">
-      <div className="flex flex-center justify-center items-center bg-grey-150 rounded-full h-12 w-12 text-violet-400">
-        <NoApplicationsForRoundIcon className="w-6 h-6" />
-      </div>
-      <NoApplicationsMessage />
-    </div>
-  );
-}
-
-function Continue(props: { onClick: () => void }) {
-  return (
-    <Button
-      type="button"
-      $variant="solid"
-      className="text-sm px-5"
-      onClick={props.onClick}
-    >
-      Continue
-    </Button>
-  );
-}
-
-function NumberOfApplicationsSelectedMessage(props: {
-  grantApplications: GrantApplication[];
-  predicate: (obj: GrantApplication) => boolean;
-}) {
-  return (
-    <span className="text-grey-400 text-sm mr-6">
-      You have selected{" "}
-      {props.grantApplications?.filter(props.predicate).length} Grant
-      Applications
-    </span>
-  );
-}
+const updateFilter = (
+  status: string,
+  selection: GrantApplication[]
+): GrantApplication[] => {
+  if (status === "PENDING") {
+    return selection?.filter((obj) => obj.status !== "PENDING");
+  } else if (status === "APPROVED") {
+    return selection?.filter((obj) => obj.status === "REJECTED");
+  } else if (status === "REJECTED") {
+    return selection?.filter((obj) => obj.status === "APPROVED");
+  } else {
+    return [];
+  }
+};
