--- packages/round-manager/src/features/round/__tests__/ApplicationsReceived.test.tsx
@@ -1,446 +1,0 @@
-/* eslint-disable @typescript-eslint/no-non-null-assertion,@typescript-eslint/no-explicit-any */
-import {
-  fireEvent,
-  render,
-  screen,
-  waitFor,
-  within,
-} from "@testing-library/react";
-import ApplicationsReceived from "../ApplicationsReceived";
-import { makeGrantApplicationData } from "../../../test-utils";
-import {
-  ApplicationContext,
-  ApplicationState,
-  initialApplicationState,
-} from "../../../context/application/ApplicationContext";
-import { MemoryRouter } from "react-router-dom";
-import {
-  BulkUpdateGrantApplicationContext,
-  initialBulkUpdateGrantApplicationState,
-} from "../../../context/application/BulkUpdateGrantApplicationContext";
-import {
-  getApplicationsByRoundId,
-  updateApplicationList,
-  updateRoundContract,
-} from "../../api/application";
-import { ApplicationStatus } from "../../api/types";
-
-jest.mock("../../api/application");
-jest.mock("../../common/Auth", () => ({
-  useWallet: () => ({
-    chain: {},
-    address: "0x0",
-    signer: {
-      getChainId: () => {
-        /* do nothing */
-      },
-    },
-    provider: { getNetwork: () => ({ chainId: "0" }) },
-  }),
-}));
-jest.mock("react-router-dom", () => ({
-  ...jest.requireActual("react-router-dom"),
-  useParams: () => ({
-    id: "some-round-id",
-  }),
-}));
-const roundIdOverride = "some-round-id";
-
-const grantApplications = [
-  makeGrantApplicationData({ roundIdOverride }),
-  makeGrantApplicationData({ roundIdOverride }),
-  makeGrantApplicationData({ roundIdOverride }),
-];
-
-grantApplications.forEach((application) => {
-  application.status = "PENDING";
-});
-
-const bulkUpdateGrantApplications = jest.fn();
-
-function setupInBulkSelectionMode() {
-  renderWithContext(<ApplicationsReceived />, {
-    applications: grantApplications,
-    isLoading: false,
-  });
-
-  const selectButton = screen.getByRole("button", {
-    name: /Select/i,
-  });
-  fireEvent.click(selectButton);
-}
-
-describe("<ApplicationsReceived />", () => {
-  beforeEach(() => {
-    (getApplicationsByRoundId as jest.Mock).mockResolvedValue(
-      grantApplications
-    );
-  });
-
-  it("should display a loading spinner if received applications are loading", () => {
-    renderWithContext(<ApplicationsReceived />, {
-      applications: [],
-      isLoading: true,
-    });
-
-    expect(screen.getByTestId("loading-spinner")).toBeInTheDocument();
-  });
-
-  describe("when there are no approved applications", () => {
-    it("should not display the bulk select option", () => {
-      renderWithContext(<ApplicationsReceived />, {
-        applications: [],
-        isLoading: false,
-      });
-
-      expect(
-        screen.queryByText(
-          "Save in gas fees by approving/rejecting multiple applications at once."
-        )
-      ).not.toBeInTheDocument();
-      expect(
-        screen.queryByRole("button", {
-          name: /Select/i,
-        })
-      ).not.toBeInTheDocument();
-    });
-  });
-
-  describe("when received applications are shown", () => {
-    it("should display the bulk select option", () => {
-      renderWithContext(<ApplicationsReceived />, {
-        applications: grantApplications,
-        isLoading: false,
-      });
-
-      expect(
-        screen.getByText(
-          "Save in gas fees by approving/rejecting multiple applications at once."
-        )
-      ).toBeInTheDocument();
-      expect(
-        screen.getByRole("button", {
-          name: /Select/i,
-        })
-      ).toBeInTheDocument();
-    });
-
-    it("should display the cancel option when select is selected", () => {
-      setupInBulkSelectionMode();
-      expect(
-        screen.getByRole("button", {
-          name: /Cancel/i,
-        })
-      ).toBeInTheDocument();
-      expect(
-        screen.queryByRole("button", {
-          name: /Select/i,
-        })
-      ).not.toBeInTheDocument();
-    });
-
-    it("should display the select option when cancel is selected", () => {
-      setupInBulkSelectionMode();
-
-      const cancelButton = screen.getByRole("button", {
-        name: /Cancel/i,
-      });
-      fireEvent.click(cancelButton);
-
-      expect(
-        screen.queryByRole("button", {
-          name: /Cancel/i,
-        })
-      ).not.toBeInTheDocument();
-      expect(
-        screen.getByRole("button", {
-          name: /Select/i,
-        })
-      ).toBeInTheDocument();
-    });
-  });
-
-  it("renders no cards when there are no projects", () => {
-    renderWithContext(<ApplicationsReceived />, {
-      applications: [],
-      isLoading: false,
-    });
-
-    expect(screen.queryAllByTestId("application-card")).toHaveLength(0);
-  });
-
-  it("renders a card for every project with PENDING status", () => {
-    renderWithContext(<ApplicationsReceived />, {
-      applications: grantApplications,
-      isLoading: false,
-    });
-
-    expect(screen.getAllByTestId("application-card")).toHaveLength(3);
-    screen.getByText(grantApplications[0].project!.title);
-    screen.getByText(grantApplications[0].project!.description);
-    screen.getByText(grantApplications[1].project!.title);
-    screen.getByText(grantApplications[1].project!.description);
-    screen.getByText(grantApplications[2].project!.title);
-    screen.getByText(grantApplications[2].project!.description);
-  });
-
-  describe("when choosing select", () => {
-    it("renders approve and reject options on each project", () => {
-      setupInBulkSelectionMode();
-
-      expect(
-        screen.queryAllByTestId("bulk-approve-reject-buttons")
-      ).toHaveLength(grantApplications.length);
-    });
-
-    it("displays an approved button as selected when approve button is selected", () => {
-      setupInBulkSelectionMode();
-
-      const approveButton = screen.queryAllByTestId("approve-button")[0];
-
-      fireEvent.click(approveButton);
-
-      expect(approveButton).toHaveClass("bg-teal-400 text-grey-500");
-    });
-
-    it("displays a rejected option as selected when reject option is selected", () => {
-      setupInBulkSelectionMode();
-
-      const rejectButton = screen.queryAllByTestId("reject-button")[0];
-      fireEvent.click(rejectButton);
-
-      expect(rejectButton).toHaveClass("bg-white text-pink-500");
-    });
-
-    describe("and when an approve option is already selected", () => {
-      it("selects the reject option and unselects the approve option when the reject option selected", () => {
-        setupInBulkSelectionMode();
-
-        const approveButton = screen.queryAllByTestId("approve-button")[0];
-        const rejectButton = screen.queryAllByTestId("reject-button")[0];
-
-        fireEvent.click(approveButton);
-        fireEvent.click(rejectButton);
-
-        expect(approveButton).not.toHaveClass("bg-teal-400 text-grey-500");
-        expect(rejectButton).toHaveClass("bg-white text-pink-500");
-      });
-
-      it("unselects the approve option when that selected approve option is selected", () => {
-        setupInBulkSelectionMode();
-
-        const approveButton = screen.queryAllByTestId("approve-button")[0];
-
-        fireEvent.click(approveButton);
-        fireEvent.click(approveButton);
-
-        expect(approveButton).not.toHaveClass("bg-teal-400 text-grey-500");
-      });
-    });
-
-    describe("and when an reject option is already selected", () => {
-      it("selects the approve button and unselects the reject button when the approve button is selected", () => {
-        setupInBulkSelectionMode();
-
-        const approveButton = screen.queryAllByTestId("approve-button")[0];
-        const rejectButton = screen.queryAllByTestId("reject-button")[0];
-
-        fireEvent.click(rejectButton);
-        fireEvent.click(approveButton);
-
-        expect(approveButton).toHaveClass("bg-teal-400 text-grey-500");
-        expect(rejectButton).not.toHaveClass("bg-white text-pink-500");
-      });
-      it("unselects the reject option when that selected reject option is selected", () => {
-        setupInBulkSelectionMode();
-
-        const rejectButton = screen.queryAllByTestId("reject-button")[0];
-
-        fireEvent.click(rejectButton);
-        fireEvent.click(rejectButton);
-
-        expect(rejectButton).not.toHaveClass("bg-white text-pink-500");
-      });
-    });
-
-    it("should approve individual applications independently", () => {
-      setupInBulkSelectionMode();
-
-      const firstApproveButton = screen.queryAllByTestId("approve-button")[0];
-      fireEvent.click(firstApproveButton);
-      expect(firstApproveButton).toHaveClass("bg-teal-400 text-grey-500");
-
-      const secondApproveButton = screen.queryAllByTestId("approve-button")[1];
-      fireEvent.click(secondApproveButton);
-      expect(secondApproveButton).toHaveClass("bg-teal-400 text-grey-500");
-    });
-
-    describe("when at least one application is selected", () => {
-      it("displays the continue option and copy", () => {
-        setupInBulkSelectionMode();
-
-        const approveButton = screen.queryAllByTestId("approve-button")[0];
-        fireEvent.click(approveButton);
-
-        const continueButton = screen.getByRole("button", {
-          name: /Continue/i,
-        });
-        expect(continueButton).toBeInTheDocument();
-        expect(
-          screen.getByText(/You have selected 1 Grant Applications/i)
-        ).toBeInTheDocument();
-
-        const approveButton2 = screen.queryAllByTestId("approve-button")[1];
-        fireEvent.click(approveButton2);
-
-        expect(continueButton).toBeInTheDocument();
-        expect(
-          screen.getByText(/You have selected 2 Grant Applications/i)
-        ).toBeInTheDocument();
-      });
-
-      it("opens confirmation when continue is selected", async () => {
-        setupInBulkSelectionMode();
-
-        const approveButton = screen.queryAllByTestId("approve-button")[0];
-        fireEvent.click(approveButton);
-
-        const continueButton = screen.getByRole("button", {
-          name: /Continue/i,
-        });
-        fireEvent.click(continueButton);
-
-        expect(screen.getByTestId("confirm-modal")).toBeInTheDocument();
-      });
-
-      it("shows the correct number of approved and rejected applications in the confirmation modal", async () => {
-        setupInBulkSelectionMode();
-
-        fireEvent.click(screen.queryAllByTestId("approve-button")[0]);
-        fireEvent.click(screen.queryAllByTestId("reject-button")[1]);
-        fireEvent.click(screen.queryAllByTestId("approve-button")[2]);
-
-        const continueButton = screen.getByRole("button", {
-          name: /Continue/i,
-        });
-        fireEvent.click(continueButton);
-
-        const approvedApplicationsCount = screen.getByTestId(
-          "approved-applications-count"
-        );
-        const rejectedApplicationsCount = screen.getByTestId(
-          "rejected-applications-count"
-        );
-
-        within(approvedApplicationsCount).getByText(/2/);
-        within(rejectedApplicationsCount).getByText(/1/);
-      });
-
-      it("starts the bulk update process when confirm is selected", async () => {
-        (updateApplicationList as jest.Mock).mockResolvedValue("");
-        (updateRoundContract as jest.Mock).mockResolvedValue("");
-        setupInBulkSelectionMode();
-
-        const approveButton = screen.queryAllByTestId("approve-button")[0];
-        fireEvent.click(approveButton);
-
-        const continueButton = screen.getByRole("button", {
-          name: /Continue/i,
-        });
-        fireEvent.click(continueButton);
-
-        const confirmationModalConfirmButton = screen.getByRole("button", {
-          name: /Confirm/i,
-        });
-
-        fireEvent.click(confirmationModalConfirmButton);
-
-        await waitFor(() => {
-          expect(updateApplicationList).toBeCalled();
-        });
-
-        await waitFor(() => {
-          expect(updateRoundContract).toBeCalled();
-        });
-
-        const updateApplicationListFirstCall = (
-          updateApplicationList as jest.Mock
-        ).mock.calls[0];
-        const actualApplicationsUpdated = updateApplicationListFirstCall[0];
-        expect(actualApplicationsUpdated).toEqual([
-          {
-            id: grantApplications[0].id,
-            round: grantApplications[0].round,
-            recipient: grantApplications[0].recipient,
-            projectsMetaPtr: grantApplications[0].projectsMetaPtr,
-            status: ApplicationStatus.APPROVED,
-          },
-        ]);
-
-        expect(updateRoundContract).toBeCalled();
-        const updateRoundContractFirstCall = (updateRoundContract as jest.Mock)
-          .mock.calls[0];
-        const actualRoundId = updateRoundContractFirstCall[0];
-        expect(actualRoundId).toEqual(roundIdOverride);
-      });
-
-      it("closes confirmation when cancel is selected", async () => {
-        setupInBulkSelectionMode();
-
-        const approveButton = screen.queryAllByTestId("approve-button")[0];
-        fireEvent.click(approveButton);
-
-        const continueButton = screen.getByRole("button", {
-          name: /Continue/i,
-        });
-        fireEvent.click(continueButton);
-
-        const modal = screen.getByTestId("confirm-modal");
-
-        const modalCancelButton = within(modal).getByRole("button", {
-          name: /Cancel/i,
-        });
-        fireEvent.click(modalCancelButton);
-
-        expect(bulkUpdateGrantApplications).not.toBeCalled();
-
-        expect(screen.queryByTestId("confirm-modal")).not.toBeInTheDocument();
-      });
-    });
-  });
-
-  describe("when bulkSelect is false", () => {
-    it("does not render approve and reject options on each card", () => {
-      renderWithContext(<ApplicationsReceived />);
-      expect(
-        screen.queryAllByTestId("bulk-approve-reject-buttons")
-      ).toHaveLength(0);
-    });
-  });
-});
-
-export const renderWithContext = (
-  ui: JSX.Element,
-  grantApplicationStateOverrides: Partial<ApplicationState> = {},
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  dispatch: any = jest.fn()
-) =>
-  render(
-    <MemoryRouter>
-      <BulkUpdateGrantApplicationContext.Provider
-        value={initialBulkUpdateGrantApplicationState}
-      >
-        <ApplicationContext.Provider
-          value={{
-            state: {
-              ...initialApplicationState,
-              ...grantApplicationStateOverrides,
-            },
-            dispatch,
-          }}
-        >
-          {ui}
-        </ApplicationContext.Provider>
-      </BulkUpdateGrantApplicationContext.Provider>
-    </MemoryRouter>
-  );
